add call_subnet and clean up plugin
