+++ plugins/bittensor/pyproject.toml
@@ -1,0 +1,30 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "bittensor_game_sdk"
+version = "0.1.0"
+authors = [{ name = "BitMind Ai" }]
+description = "Official Bittensor Python SDK for GAME by Virtuals"
+readme = "README.md"
+requires-python = ">=3.9"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+    "Development Status :: 3 - Alpha",
+    "Intended Audience :: Developers",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+]
+dependencies = ["game-sdk>=0.1.1", "requests>=2.31.0", "aiohttp>=3.8.6"]
+
+[project.urls]
+"Homepage" = "https://github.com/game-by-virtuals/game-python/plugins/allora"
+"Bug Tracker" = "https://github.com/game-by-virtuals/game-python"
+
+[tool.hatch.build.targets.wheel]
+packages = ["bittensor_game_sdk"]
